+++ apps/web/app/future/auth/sso/[provider]/page.tsx
@@ -1,12 +1,14 @@
-import Provider from "@pages/auth/sso/[provider]";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { WithLayout } from "app/layoutHOC";
-import type { InferGetServerSidePropsType } from "next";
+
+import Provider from "@components/pages/auth/sso/[provider]";
 
 import { getServerSideProps } from "@server/lib/auth/sso/[provider]/getServerSideProps";
+
+const getData = withAppDirSsr(getServerSideProps);
 
 export default WithLayout({
   getLayout: null,
   Page: Provider,
-  getData: withAppDirSsr<InferGetServerSidePropsType<typeof getServerSideProps>>(getServerSideProps),
+  getData,
 })<"P">;
